requirements.txt
accelerate==1.7.0
huggingface-hub==0.32.4
transformers==4.52.4
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
safetensors==0.5.3
tqdm==4.67.1
matplotlib==3.10.3
numpy==2.2.6
Pillow==11.2.1
requests==2.32.3
regex==2024.11.6
sympy==1.14.0
networkx==3.4.2
fsspec==2025.5.1
